+++ Changelog.md
@@ -7,6 +7,8 @@
 
 - Added the `priceGreaterThan`, `priceLessThan` and `priceBetween` methods to the ProductSearch class by [Matima](https://github.com/mahdirezaei-dev) in [#176](https://github.com/vanilophp/framework/pull/176)
 - Added the `Macroable` trait to the `ProductSearch` class
+- Added the possibility to retrieve the link items directly using `linkItems()` method as `Get::the($type)->linkItems()->of($model)`
+- Added the `link_items` helper (shortcut to Get::the()->linkItems()
 - Changed the offline payment gateway's icon from a circle to a plug+x
 
 ## 4.0.1